fixed import errors and ditto
