+++ App/views/__init__.py
@@ -6,8 +6,8 @@ from .index import index_views
 from .login import login_views
 from .signup import signup_views
 from .user import user_views
-from .user_search import user_search_views
 from .user_history import user_history_views
+from .user_search import user_search_views
 
 
 views = [
@@ -17,6 +17,6 @@ views = [
     login_views,
     signup_views,
     user_views,
-    user_search_views, 
-    user_history_views]
+    user_history_views,
+    user_search_views]
 # blueprints must be added to this list
